requirements.txt
streamlit==1.20.0
langchain==0.0.163
langchain-groq==0.0.1
langchain-community==0.0.1
faiss-cpu==1.7.2
huggingface-hub==0.15.1
openai==0.27.0
python-dotenv==0.21.1
